Api Lab- Start of AddressController

+++ lab-07-orm-Queries/src/main/java/com/cydeo/entity/Payment.java
@@ -7,13 +7,14 @@ import lombok.NoArgsConstructor;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
+import java.math.BigDecimal;
 
 @Entity
 @Data
 @NoArgsConstructor
 public class Payment extends BaseEntity{
 
-    private Double paidPrice;
+    private BigDecimal paidPrice;
 
     @Enumerated(EnumType.STRING)
     private PaymentMethod paymentMethod;
